refacgtor GatewayManager
