+++ src/lib/nodes/CargoMessageStream.ts
@@ -1,0 +1,4 @@
+export type CargoMessageStream = AsyncIterable<{
+  readonly message: Buffer;
+  readonly expiryDate: Date;
+}>;
